Add files via upload
This project is intended for educational purposes only. It is provided "as-is" without any warranties or guarantees. The author takes no responsibility for any misuse or legal liabilities resulting from the use of this project in other contexts.

By using this code, you agree that you are solely responsible for compliance with any laws or regulations that may apply.

Use at your own risk.